workround for print warning

+++ src/CascadedKernels.cuh
@@ -489,7 +489,7 @@ __device__ void block_bitpack(
 
   if (threadIdx.x == 0) {
     printf("bitwidth %u frame_of_reference %u num_elements %u\n",
-           bitwidth, frame_of_reference, num_elements);
+           bitwidth, (uint32_t)frame_of_reference, (uint32_t)num_elements);
   }
 
   current_ptr = reinterpret_cast<uint32_t*>(
