+++ src/nvcomp_common_deps/hlif_shared.cuh
@@ -163,7 +163,8 @@ __device__ inline void HlifCompressBatch(
   int initial_chunks = gridDim.x * chunks_per_block;
 
   if (cg_group.thread_rank() == 0) {
-    printf("this_ix_chunk %u num_chunks %u\n", this_ix_chunk, compression_args.num_chunks);
+    printf("this_ix_chunk %u num_chunks %u\n",
+           this_ix_chunk, (uint32_t)compression_args.num_chunks);
   }
 
   while (this_ix_chunk < compression_args.num_chunks) {
